+++ coders/pcx.c
@@ -1069,7 +1069,7 @@ static unsigned int WritePCXImage(const ImageInfo *image_info,Image *image)
       write_dcx=MagickTrue;
       (void) WriteBlobLSBLong(image,0x3ADE68B1L);
       page_table=MagickAllocateResourceLimitedClearedArray(ExtendedSignedIntegralType *,
-                                                           max_scenes+1,
+                                                           (size_t) max_scenes+1,
                                                            sizeof(ExtendedSignedIntegralType));
       if (page_table == (ExtendedSignedIntegralType *) NULL)
         ThrowPCXWriterException(ResourceLimitError,MemoryAllocationFailed,image);
